Fix for Issue #23. Made sure log backups had the correct year. Also made sure restarting the flight computer wouldn't overwrite the log. -splice can't write to files in append mode -Set output file offset to the result of a fseek() operation 	-fix compiler errors

+++ MOSES_FC/logger.c
@@ -93,7 +93,7 @@ void copy_log_to_disk() {
         printf("%d - %s\n", errno, strerror(errno));
 
     /* Move the cursor to the end of the backup log*/
-    int out_offset = fseek(out_file, 0, SEEK_END);
+    loff_t out_offset = fseek(out_file, 0, SEEK_END);
     if (out_offset == -1)
         printf("%d - %s\n", errno, strerror(errno));
